+++ src/main/java/com/account/yomankum/security/jwt/TokenServiceImpl.java
@@ -1,9 +1,9 @@
 package com.account.yomankum.security.jwt;
 
-import com.account.yomankum.security.oauth.GoogleJwt;
-import com.account.yomankum.security.oauth.JwtValue;
-import com.account.yomankum.security.oauth.KakaoJwt;
-import com.account.yomankum.security.oauth.Sns;
+import com.account.yomankum.security.domain.GoogleJwt;
+import com.account.yomankum.security.domain.JwtValue;
+import com.account.yomankum.security.domain.KakaoJwt;
+import com.account.yomankum.security.domain.Sns;
 import com.account.yomankum.domain.enums.Name;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
